requirements.txt
discord.py
aiohttp
aiosqlite